chore: factor sql and gql clients out of base test case
